Development

+++ viva/font.cpp
@@ -14,7 +14,7 @@ namespace viva
         // tex: texture to use
         // letterSize: width and height of each letter in pixels
         // charsPerRow: how many characters there are per row
-        Font::Font(Texture* tex, const Size& letterSize, int charsPerRow)
+        Font::Font(Texture* tex, const Size& letterSize, uint charsPerRow)
         {
             texture = tex;
 
@@ -22,10 +22,10 @@ namespace viva
             {
                 for (int j = 0; j < charsPerRow && k < 127; j++, k++)
                 {
-                    charsUv.push_back(Rect(j*letterSize.Width / (float)texture->GetSize().Width,
-                        i*letterSize.Height / (float)texture->GetSize().Height,
-                        (j + 1)*letterSize.Width / (float)texture->GetSize().Width,
-                        (i + 1)*letterSize.Height / (float)texture->GetSize().Height));
+                    charsUv.push_back(Rect(j*letterSize.Width / texture->GetSize().Width,
+                        i*letterSize.Height / texture->GetSize().Height,
+                        (j + 1)*letterSize.Width / texture->GetSize().Width,
+                        (i + 1)*letterSize.Height / texture->GetSize().Height));
                 }
             }
         }
